Remove old implemtation for window, process and smart (#974)
