+++ src/utils/ADempiere/dictionaryUtils.js
@@ -20,7 +20,6 @@ import { getContext, getParentFields, getPreference, parseContext } from '@/util
 import REFERENCES, { BUTTON, DEFAULT_SIZE, FIELDS_HIDDEN } from '@/utils/ADempiere/references'
 import { FIELD_OPERATORS_LIST } from '@/utils/ADempiere/dataUtils'
 import language from '@/lang'
-import { config } from '@/utils/ADempiere/config'
 
 /**
  * Generate field to app
@@ -686,25 +685,16 @@ export function convertAction(action) {
       actionAttributes.name = 'process'
       actionAttributes.icon = 'component'
       actionAttributes.component = () => import('@/views/ADempiere/Process')
-      if (config.betaFunctionality.process) {
-        actionAttributes.component = () => import('@/views/ADempiere/ProcessView')
-      }
       break
     case 'R':
       actionAttributes.name = 'report'
       actionAttributes.icon = 'skill'
       actionAttributes.component = () => import('@/views/ADempiere/Process')
-      if (config.betaFunctionality.process) {
-        actionAttributes.component = () => import('@/views/ADempiere/ProcessView')
-      }
       break
     case 'S':
       actionAttributes.name = 'browser'
       actionAttributes.icon = 'search'
       actionAttributes.component = () => import('@/views/ADempiere/Browser')
-      if (config.betaFunctionality.process) {
-        actionAttributes.component = () => import('@/views/ADempiere/BrowserView')
-      }
       break
     case 'T':
       actionAttributes.name = 'task'
@@ -714,9 +704,6 @@ export function convertAction(action) {
       actionAttributes.name = 'window'
       actionAttributes.icon = 'tab'
       actionAttributes.component = () => import('@/views/ADempiere/Window')
-      if (config.betaFunctionality.window) {
-        actionAttributes.component = () => import('@/views/ADempiere/WindowView')
-      }
       break
     case 'X':
       actionAttributes.name = 'form'
