+++ src/views/ADempiere/Test/MultiTabWindow/index.vue
@@ -28,12 +28,12 @@
 <script>
 import { defineComponent } from '@vue/composition-api'
 
-import WindowView from '@/views/ADempiere/WindowView'
+import WindowView from '@/views/ADempiere/Window'
 import multiTabMetadata from './multiTabWindow.json'
 import { convertObjectToKeyValue } from '@/utils/ADempiere/valueFormat.js'
 
 export default defineComponent({
-  name: 'TestWindowView',
+  name: 'TestWindow',
 
   setup(props, { root }) {
     // Business Partner
