+++ src/views/ADempiere/Window/StandardWindow.vue
