+++ src/views/ADempiere/Window/windowUtils.js
